requirements.txt
openai>=1.0
flask>=3.0
flask-cors>=4.0
pyyaml>=6.0
python-dotenv>=1.0
requests>=2.31.0
